Create 1_21.c
Exercise 1-21. Write a program entab that replaces strings of blanks by the minimum number of tabs and blanks to achieve
the same spacing. Use the same tab stops as for detab. When either a tab or a single blank would suffice to reach a tab stop,
which should be given preference?